third_party/blink/web_tests/external/wpt/css/css-scroll-anchoring/text-anchor-in-vertical-rl-expected.txt

This is a testharness.js-based test.
FAIL Line at edge of scrollport shouldn't jump visually when content is inserted before assert_equals: expected -400 but got 300
Harness: the test ran to completion.
